Login: folder structure refactoring
