+++ client/src/components/app/Login/LoginButton/styles.module.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/_palette';
-@import '../../../styles/_variables';
+@import '../../../../styles/_palette';
+@import '../../../../styles/_variables';
 
 .button {
   background-color: $primary;
