+++ client/src/components/app/Login/LoginInput/styles.module.scss
@@ -1,5 +1,5 @@
-@import '../../../styles/_palette';
-@import '../../../styles/_variables';
+@import '../../../../styles/_palette';
+@import '../../../../styles/_variables';
 
 .container {
   input {
