+++ client/src/components/app/Login/styles.module.scss
@@ -1,5 +1,5 @@
-@import '../../styles/_palette';
-@import '../../styles/_variables';
+@import '../../../styles/_palette';
+@import '../../../styles/_variables';
 
 .section {
   &,
@@ -16,7 +16,7 @@
   }
 
   .footer {
-    background: url(../../assets/logInGrass.png);
+    background: url(../../../assets/logInGrass.png);
     bottom: 0;
     height: 176px;
     opacity: 0.7;
